requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.2
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
psycopg2-binary==2.9.5
pyrsistent==0.19.3
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.2
swagger-ui-bundle==0.0.9
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
